requirements.txt
google-generativeai
python-dotenv
openai
pygame 
flask
flask_cors
kubernetes
google-cloud-texttospeech